pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "endpulse"
version = "0.3.0"
description = "Lightweight CLI tool for API endpoint health checking and response time monitoring"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{name = "Kim M. Hinton", email = "kim.hinton00@gmail.com"}]
keywords = ["api", "health-check", "monitoring", "cli", "endpoint"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "respx>=0.21",
]

[project.scripts]
endpulse = "endpulse.cli:main"

[project.urls]
Homepage = "https://github.com/kimhinton/endpulse"
Repository = "https://github.com/kimhinton/endpulse"
Issues = "https://github.com/kimhinton/endpulse/issues"

[tool.setuptools.packages.find]
include = ["endpulse*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
strict = true

[tool.coverage.run]
source = ["endpulse"]
